api_requests/requirements.txt
mysql-connector-python
PyYAML
requests
pycountry